+++ NEWS.md
@@ -1,5 +1,8 @@
 # rules (development version)
 
+* Added `tidy()` methods for cubist and RuleFit models to get back the rules
+  and model coefficients. 
+
 # rules 0.1.0
 
 * The default encoding methods were changed such that `parsnip` will respect the encoding that the underlying model uses.